Move workflow/sinks/ → workflow/tasks/functions/sinks/
Sinks are not core workflow primitives (like variables) — they are
shared helpers used only by task plugins. workflow/tasks/functions/
is the right layer: shared utilities called by tasks, not tasks themselves.
Mirrors internal tasks/functions/data/spark/sink.py convention.

Folder taxonomy:
  workflow/schema/sinks/          — SinkResult + typed config dataclasses
  workflow/tasks/functions/sinks/ — DataSink ABC + implementations
  workflow/tasks/pusher/          — task plugins that call the above
