--- python/michelangelo/workflow/sinks/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for michelangelo.workflow.sinks."""
