+++ python/michelangelo/workflow/tasks/functions/__init__.py
@@ -1,0 +1,1 @@
+"""workflow.tasks.functions package — shared functions used by task plugins."""
