+++ python/michelangelo/workflow/tasks/functions/sinks/__init__.py
@@ -4,7 +4,7 @@ Each sink accepts a typed config dataclass from
 ``michelangelo.workflow.schema.sinks`` — validated at pipeline-definition time:
 
     from michelangelo.workflow.schema.sinks import HiveSinkConfig, LocalFileSinkConfig
-    from michelangelo.workflow.sinks import (
+    from michelangelo.workflow.tasks.functions.sinks import (
         DataSink, HiveSink, LocalFileSink, InMemorySink
     )
 
@@ -17,9 +17,9 @@ Provider layers extend this by subclassing ``DataSink`` and registering their
 own config dataclasses in ``schema/sinks/``.
 """
 
-from michelangelo.workflow.sinks.base import DataSink
-from michelangelo.workflow.sinks.hive import HiveSink
-from michelangelo.workflow.sinks.local import LocalFileSink
-from michelangelo.workflow.sinks.memory import InMemorySink
+from michelangelo.workflow.tasks.functions.sinks.base import DataSink
+from michelangelo.workflow.tasks.functions.sinks.hive import HiveSink
+from michelangelo.workflow.tasks.functions.sinks.local import LocalFileSink
+from michelangelo.workflow.tasks.functions.sinks.memory import InMemorySink
 
 __all__ = ["DataSink", "HiveSink", "InMemorySink", "LocalFileSink"]
