+++ python/michelangelo/workflow/tasks/functions/sinks/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for michelangelo.workflow.tasks.functions.sinks."""
